Initial commit
At this point the pinhole colleciton is passing almost all of the unit
tests of the individual libraries. The major obstacle at the moment is
that the paginated collection does not do intelligent updates when
adding / removing models, always doing a full reset. This is the next
thing I'll be tackling.